refactor(lakehouse-backup): share retry policy and terminal-error classification
Hoists the copier's private isTerminal beside the backoff helper so the target
listing can reuse it, and collects the attempt counts and delay in RetryPolicy.
